requirements.txt
google-cloud-pubsub==2.14.0
requests==2.26.0
python-dotenv
